:sparkles: Add support for snake_case and camelCase options for add_table, add_moc and add_overlay_from_stcs
