+++ js/widget.js
@@ -1,17 +1,9 @@
 import A from "https://esm.sh/aladin-lite@3.4.0-beta";
 import "./widget.css";
 import EventHandler from "./models/event_handler";
+import { camelCaseToSnakeCase } from "./utils";
 
 let idxView = 0;
-
-function camelCaseToSnakeCase(pyname) {
-  if (pyname.charAt(0) === "_") pyname = pyname.slice(1);
-  let temp = pyname.split("_");
-  for (let i = 1; i < temp.length; i++) {
-    temp[i] = temp[i].charAt(0).toUpperCase() + temp[i].slice(1);
-  }
-  return temp.join("");
-}
 
 function initAladinLite(model, el) {
   let initOptions = {};
